Move Application up a folder
